+++ crates/dslcad/src/library/text.rs
@@ -1,0 +1,60 @@
+use crate::runtime::{RuntimeError, Value};
+use std::collections::HashMap;
+
+pub fn add(left: String, right: String) -> Result<String, RuntimeError> {
+    Ok(left + &right)
+}
+
+pub fn string(item: Value) -> Result<String, RuntimeError> {
+    match item {
+        Value::Number(n) => Ok(n.to_string()),
+        Value::Bool(n) => Ok(n.to_string()),
+        Value::Text(n) => Ok(n),
+        _ => Err(RuntimeError::UnexpectedType(item.get_type())),
+    }
+}
+
+pub fn format(arguments: &HashMap<String, Value>) -> Result<String, RuntimeError> {
+    let message = arguments
+        .get("message")
+        .ok_or_else(|| RuntimeError::UnsetParameter("message".into()))?;
+    let mut message = message
+        .to_text()
+        .ok_or_else(|| RuntimeError::UnexpectedType(message.get_type()))?;
+
+    for (key, value) in arguments {
+        message = message.replace(&format!("{{{}}}", key), &string(value.clone())?);
+    }
+
+    Ok(message)
+}
+
+pub fn formatln(arguments: &HashMap<String, Value>) -> Result<String, RuntimeError> {
+    Ok(format(arguments)? + "\n")
+}
+
+#[cfg(test)]
+pub mod tests {
+    use super::*;
+
+    #[test]
+    fn it_formats_strings() {
+        assert_eq!(
+            "hi".to_owned(),
+            format(&HashMap::from([(
+                String::from("message"),
+                Value::Text("hi".into())
+            )]))
+            .unwrap()
+        );
+
+        assert_eq!(
+            "hi dom".to_owned(),
+            format(&HashMap::from([
+                (String::from("message"), Value::Text("hi {name}".into())),
+                (String::from("name"), Value::Text("dom".into()))
+            ]))
+            .unwrap()
+        );
+    }
+}